Refactor: Update data models and repository logic for Firestore compatibility
This commit introduces several changes to align data models and repository logic with Firestore, including:

- **Timestamp Handling:**
    - Standardized timestamp representation to `Long` (milliseconds since epoch) across models (`Session`, `TrackPoint`, `User`).
    - Added `nowEpochMillis()` utility function.
    - Implemented `toTimestamp()` and `toEpochMillis()` extension functions for converting between `Long` and `dev.gitlive.firebase.firestore.Timestamp`.
    - Updated `RemoteFirebaseRepository` to use `FieldValue.serverTimestamp()` for `created_at` in sessions.
- **Model Updates:**
    - `Session`:
        - Changed `id` type from `Int?` to `String`.
        - Timestamps (`startTime`, `endTime`, `createdAt`) are now nullable `Long`.
        - Default `startTime` and `createdAt` to `nowEpochMillis()`.
    - `TrackPoint`:
        - Changed `id` default to `nowEpochMillis()`.
        - Changed `sessionId` type from `Long` to `String`.
        - `timestamp` is now `Long`.
    - `User`:
        - `createdAt` is now nullable `Long` and defaults to `nowEpochMillis()`.
        - Other preference fields (`unitPreference`, `enabledNotifications`, etc.) are now nullable.
- **Repository and DAO Adjustments:**
    - `RemoteFirebaseRepository`:
        - Implemented `createSession()` to generate a new session with a default title based on the current time and save it to Firestore.
        - Updated `getSessions()` and `saveSession()` to use new Firestore mapping extensions.
    - `TrackPointsDao` and `SessionDao`:
        - Adjusted to handle `String` `sessionId`.
    - `RemoteSessionsRepository`:
        - Ensured `createSession()` uses the updated `Session` model.
- **Extension Functions:**
    - Added `SessionFirestore+Extensions.kt` for `toSession()` (Firestore `DocumentSnapshot` to `Session`) and `toFirestoreMap()` (`Session` to Firestore `Map`).
    - Added `TrackPointFirestore+Extensions.kt` for `toTrackPoint()` and `toFirestoreMap()`.
    - Updated `TrackPointEntity+Extensions.kt` and `SessionEntity+Extensions.kt` for SQLDelight entity mapping, including `toEntity()` methods.
    - Added `Instant+Extentions.kt` with `asSessionTitle()` for generating default session titles.
- **UI Adjustments:**
    - `SettingsScreen`: Handle nullable user preferences with default values.
    - `HistoryScreen`: Updated dummy data to reflect `String` session IDs and nullable timestamps.
    - `SessionItem`: Adapted to nullable `Long` timestamps.
- **Dependency Management:**
    - Added `firebase-firestore-ktx` to `libs.versions.toml` and `shared/build.gradle.kts`.
- **Code Cleanup:**
    - Removed unused utility files (`HaversineUtil.kt`, `TrackPointEntityExtension.kt`, `UserEntityExtension.kt`, `SessionEntityExtension.kt`).
    - Corrected package for `StopSession.kt`.

+++ shared/src/commonMain/kotlin/com/topout/kmp/data/dao/TrackPointsDao.kt
@@ -30,7 +30,7 @@ class TrackPointsDao(
         metrics: Metrics
     ): String {
         queries.insertTrackPoint(
-            sessionId = sessionId.toLong(),
+            sessionId = sessionId,
             ts = ts,
             lat = lat,
             lon = lon,
@@ -55,26 +55,22 @@ class TrackPointsDao(
     }
 
     suspend fun getTrackPointsBySessionId(sessionId: String): List<TrackPoint> {
-        return queries.getTrackPointsBySession(sessionId.toLong())
+        return queries.getTrackPointsBySession(sessionId)
             .executeAsList()
             .map { it.toTrackPoint() }
     }
 
     fun getTrackPointsFlowBySessionId(sessionId: String): Flow<List<TrackPoint>> {
-        return queries.getTrackPointsBySession(sessionId.toLong())
+        return queries.getTrackPointsBySession(sessionId)
             .asFlow()
             .mapToList(kotlinx.coroutines.Dispatchers.IO)
             .map { list -> list.map { it.toTrackPoint() } }
     }
 
     suspend fun deleteTrackPointsBySessionId(sessionId: String) {
-        queries.deleteTrackPointsBySession(sessionId.toLong())
+        queries.deleteTrackPointsBySession(sessionId)
     }
 
     suspend fun endCurrentSession() {
-        // This method can be used to perform any cleanup or finalization
-        // For example, you might want to delete all track points for the current session
-        // or perform any other necessary operations before ending the session.
-        // Implementation depends on your specific requirements.
     }
 }
